// src/app/suites/models/suite.model.ts
export interface Suite {
    key?: any;
    id?: number;
    name?: string;
    type?: string;
    size?: string;
    area?: number;
    clients?: Array<any>;
    categories?: Array<any>;
    active?: boolean;
    note?: string;
    
    
}

// export class Category {
//     id: string;
//     title: string;
//     active: boolean;
//     required: boolean;
//     sub_items: Array<String>;
//     price: number;
//     cost: number;
//     option: string;
//     note: string;

// }